+++ include/unity/scopes/internal/ScopeObjectBase.h
@@ -19,7 +19,6 @@
 #pragma once
 
 #include <unity/scopes/internal/AbstractObject.h>
-#include <unity/scopes/internal/InvokeInfo.h>
 #include <unity/scopes/internal/MWQueryCtrlProxyFwd.h>
 #include <unity/scopes/internal/MWReplyProxyFwd.h>
 #include <unity/scopes/ChildScope.h>
@@ -39,6 +38,8 @@ class CannedQuery;
 
 namespace internal
 {
+
+class InvokeInfo;
 
 class ScopeObjectBase : public AbstractObject
 {