+++ src/scopes/internal/smartscopes/SSRegistryObject.cpp
@@ -50,7 +50,7 @@ SSRegistryObject::SSRegistryObject(MiddlewareBase::SPtr middleware,
     : ssclient_(std::make_shared<SmartScopesClient>(
                     std::make_shared<HttpClientQt>(ss_config.http_reply_timeout() * 1000),  // need millisecs
                     std::make_shared<JsonCppNode>(),
-                    middleware->runtime()->logger(),
+                    middleware->runtime(),
                     sss_url))
     , refresh_stopped_(false)
     , middleware_(middleware)
@@ -82,8 +82,7 @@ SSRegistryObject::SSRegistryObject(MiddlewareBase::SPtr middleware,
     catch (std::exception const& e)
     {
 
-        BOOST_LOG_SEV(middleware_->runtime()->logger(), Logger::Error)
-            << "SSRegistryObject: get_remote_scopes() failed: " << e.what();
+        BOOST_LOG(middleware_->runtime()->logger()) << "SSRegistryObject: get_remote_scopes() failed: " << e.what();
     }
 
     refresh_thread_ = std::thread(&SSRegistryObject::refresh_thread, this);
@@ -96,7 +95,7 @@ SSRegistryObject::SSRegistryObject(MiddlewareBase::SPtr middleware,
         }
         catch (std::exception const& e)
         {
-            BOOST_LOG_SEV(middleware_->runtime()->logger(), Logger::Error)
+            BOOST_LOG(middleware_->runtime()->logger())
                 << "SSRegistryObject(): failed to create registry publisher: " << e.what();
         }
     }
@@ -210,7 +209,7 @@ void SSRegistryObject::refresh_thread()
             }
             catch (std::exception const& e)
             {
-                BOOST_LOG_SEV(middleware_->runtime()->logger(), Logger::Error)
+                BOOST_LOG(middleware_->runtime()->logger())
                     << "SSRegistryObject: get_remote_scopes() failed: " << e.what();
             }
         }
@@ -284,7 +283,7 @@ void SSRegistryObject::get_remote_scopes()
                 }
                 catch (ResourceException const& e)
                 {
-                    BOOST_LOG_SEV(middleware_->runtime()->logger(), Logger::Error)
+                    BOOST_LOG(middleware_->runtime()->logger())
                         << "SSRegistryObject: ignoring invalid settings JSON for scope \"" << scope.id << "\": "
                         << e.what();
                 }
@@ -333,7 +332,7 @@ void SSRegistryObject::get_remote_scopes()
                 }
                 catch (ResourceException const& e)
                 {
-                    BOOST_LOG_SEV(middleware_->runtime()->logger(), Logger::Error)
+                    BOOST_LOG(middleware_->runtime()->logger())
                         << "SSRegistryObject: ignoring invalid settings JSON for scope \"" << scope.id << "\": "
                         << e.what();
                 }
@@ -348,7 +347,6 @@ void SSRegistryObject::get_remote_scopes()
             metadata->set_keywords(scope.keywords);
 
             ScopeProxy proxy = ScopeImpl::create(middleware_->create_scope_proxy(scope.id, ss_scope_endpoint_),
-                                                 middleware_->runtime(),
                                                  scope.id);
 
             metadata->set_proxy(proxy);
@@ -360,7 +358,7 @@ void SSRegistryObject::get_remote_scopes()
         }
         catch (std::exception const& e)
         {
-            BOOST_LOG_SEV(middleware_->runtime()->logger(), Logger::Error)
+            BOOST_LOG(middleware_->runtime()->logger())
                 << "SSRegistryObject: skipping scope \"" << scope.id << "\": " << e.what();
         }
     }